37: mmm-persistence: create JPQL query builder 36: mmm-persistence: refactoring refactored and improved search and integrated JPQL builder

+++ mmm-util/mmm-util-core/src/main/java/net/sf/mmm/util/search/api/SearchCriteria.java
@@ -5,8 +5,15 @@ package net.sf.mmm.util.search.api;
 import java.io.Serializable;
 
 /**
- * This is the interface for the transfer-object with the criteria for a search query. Such object specifies
- * the criteria which {@link SearchResult#getHits() hits} will match when performing a search.
+ * This is the interface for a transfer-object with the criteria for a {@link SearchQuery search query}. Such
+ * object specifies the criteria selecting which {@link SearchResult#getHits() hits} will match when
+ * performing a search.<br/>
+ * <b>NOTE:</b><br/>
+ * This interface only holds generic settings for the query such as {@link #getHitOffset()},
+ * {@link #getMaximumHitCount()}, and {@link #getSearchTimeout()}. For your individual search, you can extend
+ * {@link net.sf.mmm.util.search.base.AbstractSearchCriteria} to create a java bean with all the fields for
+ * your search. For searches in a database using JPA there is additional support provided by
+ * <code>mmm-persistence</code>.
  * 
  * @see SearchResult
  * 
@@ -16,7 +23,9 @@ import java.io.Serializable;
 public interface SearchCriteria extends Serializable {
 
   /**
-   * This method gets the maximum number of hits that will be received as result for this query.
+   * This method gets the maximum number of hits that will be received as result for this query.<br/>
+   * <b>Note:</b><br/>
+   * This feature is the same as <code>Query.setMaxResults(int)</code> in JPA.
    * 
    * @return the maximum hit-count or <code>null</code> for NO limit.
    */
@@ -24,17 +33,21 @@ public interface SearchCriteria extends Serializable {
 
   /**
    * This method gets the offset for the first hit. This is <code>0</code> by default. By providing a multiple
-   * of {@link #getMaximumHitCount()} you can simply implement paging.
+   * of {@link #getMaximumHitCount()} you can simply implement paging.<br/>
+   * <b>Note:</b><br/>
+   * This feature is the same as <code>Query.setFirstResult(int)</code> in JPA.
    * 
    * @return the offset of the first hit.
    */
   int getHitOffset();
 
   /**
-   * This method gets the maximum delay the search may last until it is canceled.
+   * This method gets the maximum delay in milliseconds the search may last until it is canceled.<br/>
+   * <b>Note:</b><br/>
+   * This feature is the same as the query hint <code>"javax.persistence.query.timeout"</code> in JPA.
    * 
-   * @return the search timeout or <code>null</code> for NO timeout.
+   * @return the search timeout in milliseconds or <code>null</code> for NO timeout.
    */
-  Integer getSearchTimeout();
+  Long getSearchTimeout();
 
 }
